requirements.txt
torch
torchvision 
segmentation-models-pytorch
numpy
pandas
pillow
opencv-python